Update getting-started.rst to clarify OS X Homebrew pip install (#2876)
* Update getting-started.rst to clarify OS X Homebrew pip install
On OS X El Capitan following the current instructions fails with fatal error: 'openssl/opensslv.h' file not found on pip install -r dev-requirements.txt; this change adds to the instructions to clarify that CFLAGS and LDFLAGS may need to be set. I based this on the instructions in the Installation page.

* Address reaperhulk comments, OpenSSL must be installed.

Clarify that OpenSSL must be installed via Homebrew or MacPorts and how to build statically.

* Make Homebrew, MacPorts, OpenSSL links

* Fix doc errors and improve overall flow
Insert tip about tox needing the same flags
Make console example multi-line so that pip install is visible
Change heading levels to allow OS X sub-section